refactor: controller 역할 분리
