+++ backend/queue-service/src/main/java/com/bidket/queue/QueueServiceApplication.java
@@ -3,6 +3,7 @@ package com.bidket.queue;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.data.redis.RedisRepositoriesAutoConfiguration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication(exclude = {RedisRepositoriesAutoConfiguration.class})
